Updated bugfix for Actions dropdown
